@@ -8,38 +8,38 @@
     "url": "https://github.com/digitalbazaar/ed25519-signature-2020"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/index.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "files": [
     "lib/*.js"
   ],
-  "module": "lib/main.js",
   "dependencies": {
-    "@digitalbazaar/ed25519-verification-key-2020": "^3.0.0",
-    "base58-universal": "^1.0.0",
-    "ed25519-signature-2020-context": "^1.0.1",
+    "@digitalbazaar/ed25519-verification-key-2020": "github:digitalbazaar/ed25519-verification-key-2020#convert-to-module",
+    "base58-universal": "github:digitalbazaar/base58-universal#fix-esm",
+    "ed25519-signature-2020-context": "^1.1.0",
     "ed25519-signature-2018-context": "^1.1.0",
-    "esm": "^3.2.25",
-    "jsonld-signatures": "^9.2.0"
+    "jsonld-signatures": "^9.3.1"
   },
   "devDependencies": {
-    "@digitalbazaar/ed25519-verification-key-2018": "^3.1.1",
-    "@digitalbazaar/security-document-loader": "^1.0.0",
-    "chai": "^4.2.0",
+    "@digitalbazaar/ed25519-verification-key-2018": "github:digitalbazaar/ed25519-verification-key-2018#convert-to-module",
+    "@digitalbazaar/security-document-loader": "github:digitalbazaar/security-document-loader#convert-to-module",
+    "c8": "^7.11.2",
+    "chai": "^4.3.6",
     "cross-env": "^7.0.3",
-    "eslint": "^7.20.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.2.0",
-    "karma": "^6.3.2",
+    "eslint": "^8.14.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "eslint-plugin-unicorn": "^42.0.0",
+    "karma": "^6.3.19",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.8",
     "karma-webpack": "^5.0.0",
-    "mocha": "^8.0.1",
+    "mocha": "^9.2.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^15.1.0",
-    "webpack": "^5.37.1"
+    "webpack": "^5.72.0"
   },
   "browser": {
     "buffer": false,
@@ -51,14 +51,14 @@
   },
   "scripts": {
     "test": "npm run lint && npm run test-node && npm run test-karma",
-    "test-karma": "karma start karma.conf.js",
+    "test-karma": "karma start karma.conf.cjs",
     "test-node": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} --require test/test-mocha.js test/*.spec.js",
-    "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
-    "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=lcovonly npm run test-node",
-    "coverage-report": "nyc report",
+    "coverage": "cross-env NODE_ENV=test c8 --reporter=lcov --reporter=text-summary npm run test-node",
+    "coverage-ci": "cross-env NODE_ENV=test c8 --reporter=lcovonly --reporter=text npm run test-node",
+    "coverage-report": "c8 report",
     "lint": "eslint ."
   },
-  "nyc": {
+  "c8": {
     "exclude": [
       "test"
     ],
