@@ -3,7 +3,7 @@
  */
 import jsigs from 'jsonld-signatures';
 const {suites: {LinkedDataSignature}} = jsigs;
-const base58btc = require('base58-universal');
+import * as base58btc from 'base58-universal';
 import {
   Ed25519VerificationKey2020
 } from '@digitalbazaar/ed25519-verification-key-2020';
