@@ -1,8 +1,7 @@
 /*!
  * Copyright (c) 2020-2021 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import suiteContext from 'ed25519-signature-2020-context';
 
-// translate `main.js` to CommonJS
-require = require('esm')(module);
-module.exports = require('./main.js');
+export {Ed25519Signature2020} from './Ed25519Signature2020.js';
+export {suiteContext};
